requirements.txt
streamlit==1.30.0
google-generativeai==0.5.4
python-dotenv==1.0.0
PyPDF2==3.0.1
scikit-learn==1.3.2
pandas==2.1.3
faiss-cpu==1.7.4
markdown==3.5.1
plotly==5.17.0
python-docx==1.0.1
numpy==1.26.4
